# doc-files/kypo-training-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 2.3.6"
  title: "KYPO Training - API Reference"
host: "localhost:8083"
basePath: "/kypo-rest-training/api/v1"
...
schemes:
- "http"
- "https"
paths:
...
  /cheating-detections/exports/{cheatingDetectionId}:
    get:
      tags:
      - "Cheating detection"
      summary: "Archive cheating detection results"
      description: ""
      operationId: "archiveCheatingDetectionResults"
      produces:
      - "application/octet-stream"
      parameters:
      - name: "cheatingDetectionId"
        in: "path"
        description: "Id of cheating detection"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Cheating Detection results archived."
          schema:
            $ref: "#/definitions/TrainingInstanceArchiveDTO"
        404:
          description: "Cheating Detection not found."
          schema:
            $ref: "#/definitions/ApiError"
        409:
          description: "Cannot archive detection that is not finished."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
  /exports/training-definitions/{definitionId}:
    get:
      tags:
      - "Export Imports"
      summary: "Get exported training definitions and levels."
      description: ""
      operationId: "getExportedTrainingDefinitionAndLevels"
      produces:
      - "application/octet-stream"
      parameters:
      - name: "definitionId"
        in: "path"
        description: "Id of training definition"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Training definitions and levels found and exported."
          schema:
            $ref: "#/definitions/ExportTrainingDefinitionAndLevelsDTO"
        404:
          description: "Training definition not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /exports/training-instances/{instanceId}:
    get:
      tags:
      - "Export Imports"
      summary: "Archive training instance"
      description: ""
      operationId: "archiveTrainingInstance"
      produces:
      - "application/octet-stream"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Id of training instance"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Training instance archived."
          schema:
            $ref: "#/definitions/TrainingInstanceArchiveDTO"
        404:
          description: "Training instance not found."
          schema:
            $ref: "#/definitions/ApiError"
        409:
          description: "Cannot archive instance that is not finished."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /exports/training-instances/{instanceId}/scores:
    get:
      tags:
      - "Export Imports"
      summary: "Export training instance scores"
      description: ""
      operationId: "exportTrainingInstanceScores"
      produces:
      - "text/plain"
      parameters:
      - name: "instanceId"
        in: "path"
        description: "Id of training instance"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Training instance score exported"
          schema:
            type: "string"
        404:
          description: "Training instance not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /imports/training-definitions:
    post:
      tags:
      - "Export Imports"
      summary: "Import training definition with levels."
      description: ""
      operationId: "importTrainingDefinition"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Training definition to be imported"
        required: true
        schema:
          $ref: "#/definitions/ImportTrainingDefinitionDTO"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      responses:
        200:
          description: "Training definition imported."
          schema:
            $ref: "#/definitions/TrainingDefinitionByIdDTO"
        422:
          description: "Sum of hints penalties in imported training level is greater\
            \ than maximal score."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    description: "Obtain bearer token from OIDC provider and enter the token in the\
      \ following field. Use format 'Bearer {token}'"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  AbstractLevelDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of level."
      title:
        type: "string"
        example: "Training Level1"
        description: "Short textual description of the level."
      max_score:
        type: "integer"
        format: "int32"
        example: 20
        description: "The maximum score a participant can achieve during a level."
      snapshot_hook:
        $ref: "#/definitions/SnapshotHookDTO"
      level_type:
        type: "string"
        example: "TRAINING"
        description: "Type of the level."
        enum:
        - "INFO_LEVEL"
        - "TRAINING_LEVEL"
        - "ACCESS_LEVEL"
        - "ASSESSMENT_LEVEL"
      estimated_duration:
        type: "integer"
        format: "int32"
        example: 5
        description: "Estimated time taken by the player to resolve the level."
      training_definition:
        example: "2"
        description: "Training definition to which is this level assigned."
        $ref: "#/definitions/TrainingDefinitionDTO"
      order:
        type: "integer"
        format: "int32"
        example: 2
        description: "Order of level, starts with 0"
      minimal_possible_solve_time:
        type: "integer"
        format: "int32"
        example: 5
        description: "Minimal possible solve time (minutes) that must be taken by\
          \ the player to solve the level."
    description: "Superclass for classes TrainingLevelDTO, AccessLevelDTO, AssessmentLevelDTO\
      \ and InfoLevelDTO"
  AbstractLevelExportDTO:
    type: "object"
    properties:
      title:
        type: "string"
        example: "Training Level1"
        description: "Short textual description of the level."
      level_type:
        type: "string"
        example: "TRAINING"
        description: "Type of the level."
        enum:
        - "INFO_LEVEL"
        - "TRAINING_LEVEL"
        - "ACCESS_LEVEL"
        - "ASSESSMENT_LEVEL"
      order:
        type: "integer"
        format: "int32"
        example: 2
        description: "Order of level, starts with 0"
      estimated_duration:
        type: "integer"
        format: "int32"
        example: 5
        description: "Estimated time (minutes) taken by the player to solve the level."
      minimal_possible_solve_time:
        type: "integer"
        format: "int32"
        example: 5
        description: "Minimal possible solve time (minutes) that must be taken by\
          \ the player to solve the level."
    description: "Superclass for classes TrainingLevelExportDTO, AccessLevelExportDTO,\
      \ InfoLevelExportDTO and AssessmentLevelExportDTO"
  AbstractLevelImportDTO:
    type: "object"
    required:
    - "estimated_duration"
    - "level_type"
    - "order"
    discriminator: "level_type"
    properties:
      title:
        type: "string"
        example: "Training Level1"
        description: "Short textual description of the level."
      level_type:
        type: "string"
        example: "TRAINING_LEVEL"
        description: "Type of the level."
        enum:
        - "INFO_LEVEL"
        - "TRAINING_LEVEL"
        - "ACCESS_LEVEL"
        - "ASSESSMENT_LEVEL"
      order:
        type: "integer"
        format: "int32"
        example: 2
        description: "Order of level, starts with 0"
        minimum: 0
      estimated_duration:
        type: "integer"
        format: "int32"
        example: 5
        description: "Estimated time (minutes) taken by the player to solve the level."
        minimum: 0
      minimal_possible_solve_time:
        type: "integer"
        format: "int32"
        example: 5
        description: "Minimal possible solve time (minutes) that must be taken by\
          \ the player to solve the level."
    description: "Superclass for classes TrainingLevelImportDTO, AccessLevelImportDTO,\
      \ AssessmentLevelImportDTO and InfoLevelImportDTO"
...
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
        - "CONTINUE"
        - "SWITCHING_PROTOCOLS"
        - "PROCESSING"
        - "CHECKPOINT"
        - "OK"
        - "CREATED"
        - "ACCEPTED"
        - "NON_AUTHORITATIVE_INFORMATION"
        - "NO_CONTENT"
        - "RESET_CONTENT"
        - "PARTIAL_CONTENT"
        - "MULTI_STATUS"
        - "ALREADY_REPORTED"
        - "IM_USED"
        - "MULTIPLE_CHOICES"
        - "MOVED_PERMANENTLY"
        - "FOUND"
        - "MOVED_TEMPORARILY"
        - "SEE_OTHER"
        - "NOT_MODIFIED"
        - "USE_PROXY"
        - "TEMPORARY_REDIRECT"
        - "PERMANENT_REDIRECT"
        - "BAD_REQUEST"
        - "UNAUTHORIZED"
        - "PAYMENT_REQUIRED"
        - "FORBIDDEN"
        - "NOT_FOUND"
        - "METHOD_NOT_ALLOWED"
        - "NOT_ACCEPTABLE"
        - "PROXY_AUTHENTICATION_REQUIRED"
        - "REQUEST_TIMEOUT"
        - "CONFLICT"
        - "GONE"
        - "LENGTH_REQUIRED"
        - "PRECONDITION_FAILED"
        - "PAYLOAD_TOO_LARGE"
        - "REQUEST_ENTITY_TOO_LARGE"
        - "URI_TOO_LONG"
        - "REQUEST_URI_TOO_LONG"
        - "UNSUPPORTED_MEDIA_TYPE"
        - "REQUESTED_RANGE_NOT_SATISFIABLE"
        - "EXPECTATION_FAILED"
        - "I_AM_A_TEAPOT"
        - "INSUFFICIENT_SPACE_ON_RESOURCE"
        - "METHOD_FAILURE"
        - "DESTINATION_LOCKED"
        - "UNPROCESSABLE_ENTITY"
        - "LOCKED"
        - "FAILED_DEPENDENCY"
        - "UPGRADE_REQUIRED"
        - "PRECONDITION_REQUIRED"
        - "TOO_MANY_REQUESTS"
        - "REQUEST_HEADER_FIELDS_TOO_LARGE"
        - "UNAVAILABLE_FOR_LEGAL_REASONS"
        - "INTERNAL_SERVER_ERROR"
        - "NOT_IMPLEMENTED"
        - "BAD_GATEWAY"
        - "SERVICE_UNAVAILABLE"
        - "GATEWAY_TIMEOUT"
        - "HTTP_VERSION_NOT_SUPPORTED"
        - "VARIANT_ALSO_NEGOTIATES"
        - "INSUFFICIENT_STORAGE"
        - "LOOP_DETECTED"
        - "BANDWIDTH_LIMIT_EXCEEDED"
        - "NOT_EXTENDED"
        - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[The requested resource was not found (different for each type of\
          \ exception).]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
    description: "Superclass for classes ApiEntityError and ApiMicroserviceError"
...
  ExportTrainingDefinitionAndLevelsDTO:
    type: "object"
    properties:
      title:
        type: "string"
        example: "TrainingDefinition2"
        description: "A name of the training/game (e.g., Photo Hunter) ."
      description:
        type: "string"
        example: "Unreleased training definition"
        description: "Description of training definition that is visible to the participant."
      prerequisites:
        type: "array"
        description: "List of knowledge and skills necessary to complete the training."
        items:
          type: "string"
      outcomes:
        type: "array"
        description: "A list of knowledge and skills that the participant should learn\
          \ by attending the training (if it is used for educational purposes) "
        items:
          type: "string"
      state:
        type: "string"
        example: "UNRELEASED"
        description: "Current state of training definition."
        enum:
        - "PRIVATED"
        - "RELEASED"
        - "ARCHIVED"
        - "UNRELEASED"
      show_stepper_bar:
        type: "boolean"
        example: false
        description: "Sign if stepper bar should be displayed."
      levels:
        type: "array"
        description: "Information about all levels in training definition."
        items:
          $ref: "#/definitions/AbstractLevelExportDTO"
      estimated_duration:
        type: "integer"
        format: "int32"
        example: 5
        description: "Estimated time (minutes) taken by the player to finish run created\
          \ from this definition."
      variant_sandboxes:
        type: "boolean"
        example: false
        description: "Marking if levels flags/answers are randomly generated and are\
          \ different for each trainee. Default is false."
    description: "An exported detailed information about training definition which\
      \ also include individual levels."
...
  ImportTrainingDefinitionDTO:
    type: "object"
    required:
    - "show_stepper_bar"
    - "state"
    properties:
      title:
        type: "string"
        example: "TrainingDefinition2"
        description: "A name of the training/game (e.g., Photo Hunter) ."
      description:
        type: "string"
        example: "Unreleased training definition"
        description: "Description of training definition that is visible to the participant."
      prerequisites:
        type: "array"
        description: "List of knowledge and skills necessary to complete the training."
        items:
          type: "string"
      outcomes:
        type: "array"
        description: "A list of knowledge and skills that the participant should learn\
          \ by attending the training (if it is used for educational purposes) "
        items:
          type: "string"
      state:
        type: "string"
        example: "UNRELEASED"
        description: "Current state of training definition."
        enum:
        - "PRIVATED"
        - "RELEASED"
        - "ARCHIVED"
        - "UNRELEASED"
      show_stepper_bar:
        type: "boolean"
        example: false
        description: "Sign if stepper bar should be displayed."
      levels:
        type: "array"
        description: "Information about all levels in training definition."
        items:
          $ref: "#/definitions/AbstractLevelImportDTO"
      estimated_duration:
        type: "integer"
        format: "int32"
        example: 5
        description: "Estimated time it takes to finish runs created from this definition."
      variant_sandboxes:
        type: "boolean"
        example: false
        description: "Marking if levels flags/answers are randomly generated and are\
          \ different for each trainee. Default is false."
    description: "A basic information about hint."
...
  SnapshotHookDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 8
        description: "Main identifier of snapshotHook."
      snapshot:
        type: "string"
        example: "Snapshot 1"
        description: "Content of snapshot. What should be executed to get to particular\
          \ state."
    description: "."
...
  TrainingDefinitionByIdDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training definition."
      title:
        type: "string"
        example: "TrainingDefinition2"
        description: "A name of the training/game (e.g., Photo Hunter) ."
      description:
        type: "string"
        example: "Unreleased training definition"
        description: "Description of training definition that is visible to the participant."
      prerequisites:
        type: "array"
        description: "List of knowledge and skills necessary to complete the training."
        items:
          type: "string"
      outcomes:
        type: "array"
        description: "A list of knowledge and skills that the participant should learn\
          \ by attending the training (if it is used for educational purposes) "
        items:
          type: "string"
      state:
        type: "string"
        example: "UNRELEASED"
        description: "Current state of training definition."
        enum:
        - "PRIVATED"
        - "RELEASED"
        - "ARCHIVED"
        - "UNRELEASED"
      beta_testing_group_id:
        type: "integer"
        format: "int64"
        example: 2
        description: "Group of organizers who is allowed to see the training definition."
      levels:
        type: "array"
        description: "Information about all levels in training definition."
        items:
          $ref: "#/definitions/AbstractLevelDTO"
      show_stepper_bar:
        type: "boolean"
        example: false
        description: "Sign if stepper bar should be displayed."
      can_be_archived:
        type: "boolean"
        example: true
        description: "Sign if training definition can be archived or not."
      estimated_duration:
        type: "integer"
        format: "int64"
        example: 5
        description: "Estimated time it takes to finish runs created from this definition."
      last_edited:
        type: "string"
        format: "date-time"
        example: "2017-10-19 10:23:54+02"
        description: "Time of last edit done to definition."
      last_edited_by:
        type: "string"
        example: "John Doe"
        description: "Name of the user who has done the last edit in definition."
      has_reference_solution:
        type: "boolean"
        description: "Indicates if any of the training levels have defined reference\
          \ solution."
    description: "A blueprint of abstract levels."
...
  TrainingDefinitionDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training definition."
      title:
        type: "string"
        example: "TrainingDefinition2"
        description: "A name of the training/game (e.g., Photo Hunter) ."
      description:
        type: "string"
        example: "Unreleased training definition"
        description: "Description of training definition that is visible to the participant."
      prerequisites:
        type: "array"
        description: "List of knowledge and skills necessary to complete the training."
        items:
          type: "string"
      outcomes:
        type: "array"
        description: "A list of knowledge and skills that the participant should learn\
          \ by attending the training (if it is used for educational purposes) "
        items:
          type: "string"
      state:
        type: "string"
        example: "UNRELEASED"
        description: "Current state of training definition."
        enum:
        - "PRIVATED"
        - "RELEASED"
        - "ARCHIVED"
        - "UNRELEASED"
      beta_testing_group_id:
        type: "integer"
        format: "int64"
        example: 14
        description: "Group of organizers who is allowed to see the training definition."
      show_stepper_bar:
        type: "boolean"
        example: false
        description: "Sign if stepper bar should be displayed."
      can_be_archived:
        type: "boolean"
        example: false
        description: "Sign if training definition can be archived or not."
      estimated_duration:
        type: "integer"
        format: "int64"
        example: 5
        description: "Estimated time it takes to finish runs created from this definition."
      last_edited:
        type: "string"
        format: "date-time"
        example: "2017-10-19 10:23:54+02"
        description: "Time of last edit done to definition."
      last_edited_by:
        type: "string"
        example: "John Doe"
        description: "Name of the user who has done the last edit in definition."
    description: "A blueprint of abstract levels."
...
  TrainingInstanceArchiveDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training instance."
      definition_id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of training definition associated with this\
          \ instance."
      start_time:
        type: "string"
        format: "date-time"
        example: "2016-10-19 10:23:54+02"
        description: "Date when training instance starts."
      end_time:
        type: "string"
        format: "date-time"
        example: "2017-10-19 10:23:54+02"
        description: "Date when training instance ends."
      title:
        type: "string"
        example: "Concluded Instance"
        description: "Short textual description of the training instance."
      organizers_ref_ids:
        type: "array"
        description: "Reference to organizersRefIds which organize training instance."
        uniqueItems: true
        items:
          type: "integer"
          format: "int64"
      access_token:
        type: "string"
        example: "pass-1234"
        description: "Token needed to access runs created from this definition"
      local_environment:
        type: "boolean"
        example: true
        description: "Indicates if local sandboxes are used for training runs."
      backward_mode:
        type: "boolean"
        example: true
        description: "Indicates if trainee can during training run move to the previous\
          \ already solved levels."
    description: "The finished and archived instance of training definition which\
      \ includes individual finished training runs of participants."
...